% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRA_invasive.R
\docType{package}
\name{eppo_EPRA_invasive}
\alias{eppo_EPRA_invasive}
\alias{eppo_EPRA_invasive-package}
\title{EPPO Express Pest Risk Analysis Reports}
\description{
Template for creating an R markdown document using the EPPO Express
Pest Risk Analysis reports
}
\details{
Based on the
\href{https://www.eppo.int/RESOURCES/eppo_standards/pm5_pra}{EPPO PRA}.
}
\section{eppo_invasive_pdf}{

Creates an R Markdown Express PRA for invasive plants template as a PDF
}

\section{eppo_invasive_html}{

Creates an R Markdown Express PRA for invasive plants template as a html
}

