% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pestrPRA.R
\encoding{UTF-8}
\name{invasive_plants_PRA1}
\alias{invasive_plants_PRA1}
\alias{eppo_invasive_pdf}
\alias{eppo_invasive_html}
\title{EPPO Express Pest Risk Analysis Reports}
\usage{
invasive_plants_PRA1(link_citations = TRUE, toc = FALSE,
  toc_depth = 3, smart = TRUE, graphics = TRUE,
  comapct_title = TRUE, self_contained = TRUE,
  always_allow_html = TRUE, urlcolor = "Cerulean",
  linkcolor = "TealBlue", citecolor = "TealBlue", highlight = "none",
  extra_dependencies = NULL, css = NULL, includes = NULL,
  keep_md = FALSE, lib_dir = NULL, md_extensions = NULL,
  pandoc_args = NULL, ...)

eppo_invasive_pdf(toc = FALSE, highlight = "default", ...)

eppo_invasive_html(toc = FALSE, highlight = "default", ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{smart}{Produce typographically correct output, converting straight
quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
ellipses.}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
its size).}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
"haddock", and "textmate". Pass \code{NULL} to prevent syntax highlighting.}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown HTML output formatter}

\item{css}{One or more css files to include}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{lib_dir}{Directory to copy dependent HTML libraries (e.g. jquery,
bootstrap, etc.) into. By default this will be the name of the document with
\code{_files} appended to it.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown HTML output formatter}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown HTML output formatter}
}
\description{
Template for creating an R markdown document using the EPPO Express
Pest Risk Analysis reports
}
\details{
Based on the
\href{https://www.eppo.int/RESOURCES/eppo_standards/pm5_pra}{EPPO PRA}.
}
