% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pestrPRAtemplates.R
\docType{package}
\name{pestrPRAtemplates}
\alias{pestrPRAtemplates}
\alias{pestrPRAtemplates-package}
\title{pestrPRAtemplates: A package for creating Express PRA Templates}
\description{
Based on the
\href{https://www.eppo.int/RESOURCES/eppo_standards/pm5_pra}{EPPO PRA}.
}
