% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pestrPRA.R
\encoding{UTF-8}
\name{eppo_invasive_pdf}
\alias{eppo_invasive_pdf}
\alias{eppo_invasive_html}
\alias{eppo_pest_pdf}
\title{EPPO Express Pest Risk Analysis Reports}
\usage{
eppo_invasive_pdf(toc = FALSE, highlight = "default", ...)

eppo_invasive_html(toc = FALSE, highlight = "default", ...)

eppo_pest_pdf(toc = FALSE, highlight = "default", ...)

eppo_pest_pdf(toc = FALSE, highlight = "default", ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown pdf output formatter}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown HTML output formatter}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown pdf output formatter}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown HTML output formatter}
}
\description{
Template for creating an R markdown document using the EPPO Express
Pest Risk Analysis reports
}
\details{
Based on the
\href{https://www.eppo.int/RESOURCES/eppo_standards/pm5_pra}{EPPO PRA}.
}
