% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pestrPRA.R
\encoding{UTF-8}
\name{eppo_invasive_pdf}
\alias{eppo_invasive_pdf}
\alias{eppo_pest_pdf}
\title{EPPO Express Pest Risk Analysis Reports}
\usage{
eppo_invasive_pdf(toc = FALSE, highlight = "default", ...)

eppo_pest_pdf(toc = FALSE, highlight = "default", ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown pdf output formatter}

\item{extra_dependencies, ...}{Additional function arguments to pass to the
base R Markdown pdf output formatter}
}
\description{
Templates for creating an R markdown document using the EPPO Express
Pest Risk Analysis (PRA) reports. \code{eppo_invasive_pdf} returns template
for invasive plant species PRA (that can be compiled with bookdown to both
html and pdf).
\code{eppo_pest_pdf} returns template for common pest PRA (that can be
compiled with bookdown to both html and pdf).
}
\details{
Based on the
\href{https://www.eppo.int/RESOURCES/eppo_standards/pm5_pra}{EPPO PRA}.
}
